refactor: extract git ui backend module
